requirements.txt
boto3>=1.30
pyyaml>=6.0
